Somewhat functional player shops
You can open a players shop and see what they are selling and what they are charging with the option of buy or close the GUI. There is currently no function to allow the player to actually buy the items yet. This will be done later when I have the time.